+++ evolve_keyboard_layout_wasm/src/lib.rs
@@ -12,6 +12,7 @@ use keyboard_layout::{
 };
 
 use layout_evaluation::{
+    cache::Cache,
     config::EvaluationParameters,
     evaluation::Evaluator,
     ngram_mapper::on_demand_ngram_mapper::OnDemandNgramMapper,
@@ -19,7 +20,7 @@ use layout_evaluation::{
     results::EvaluationResult,
 };
 
-use layout_optimization::common::{Cache, PermutationLayoutGenerator};
+use layout_optimization::common::PermutationLayoutGenerator;
 use layout_optimization_genevo::optimization as genevo_optimization;
 use layout_optimization_sa::optimization as sa_optimization;
