+++ layout_evaluation/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod cache;
 pub mod config;
 pub mod evaluation;
 pub mod metrics;
