+++ layout_optimization_abc/src/optimization.rs
@@ -1,7 +1,7 @@
 use keyboard_layout::{layout::Layout, layout_generator::NeoLayoutGenerator};
-use layout_evaluation::evaluation::Evaluator;
+use layout_evaluation::{cache::Cache, evaluation::Evaluator};
 
-use layout_optimization::common::{Cache, PermutationLayoutGenerator};
+use layout_optimization::common::PermutationLayoutGenerator;
 
 use anyhow::Result;
 use serde::Deserialize;
